Be sign correct.

+++ crawl-ref/source/describe-spells.cc
@@ -321,9 +321,9 @@ vector<spell_type> map_chars_to_spells(const spellset &spells,
         return flat_spells;
 
     vector<spell_type> column_spells;
-    for (int i = 0; i < flat_spells.size(); i += 2)
+    for (size_t i = 0; i < flat_spells.size(); i += 2)
         column_spells.emplace_back(flat_spells[i]);
-    for (int i = 1; i < flat_spells.size(); i += 2)
+    for (size_t i = 1; i < flat_spells.size(); i += 2)
         column_spells.emplace_back(flat_spells[i]);
     return column_spells;
 }
